Moved EditPOIDialog to seperate module
